web/app/requirements.txt
panel>=1.3.0
plotly>=5.0.0
pandas>=1.3.0
requests>=2.25.0
python-dotenv>=0.19.0
matplotlib>=3.5.0
param>=1.12.0
